+++ packages/visual-dev/CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - Fixed a visual regression with the upwards menu direction where there was an unnecessary gap between the handle and list
   - \<RadioCard>
     - Added more flexibility to allow custom CSS and better design for cards without icons
+    - Added a disabled state
 
 ## [1.2.0] - 2022-11-29